Move class from logistics to model. Change protected ctor to public in model.customers.
