+++ customers/src/main/java/com/alevel/deliverit/customers/LogisticsGateway.java
@@ -1,0 +1,13 @@
+package com.alevel.deliverit.customers;
+
+import com.alevel.deliverit.logistics.postal.network.Route;
+
+/**
+ * @author Sergey Bogovesov
+ */
+public class LogisticsGateway {
+
+    public static Route findRoute(RouteLookupRequest request) {
+        return null;
+    }
+}
