+++ customers/src/main/java/com/alevel/deliverit/customers/ParcelReception.java
@@ -6,6 +6,7 @@ import com.alevel.deliverit.TrackNumbers;
 import com.alevel.deliverit.billing.Money;
 import com.alevel.deliverit.logistics.EstimatedDeliveryTime;
 import com.alevel.deliverit.logistics.TrackNumber;
+import com.alevel.deliverit.logistics.postal.network.Route;
 import com.google.common.annotations.VisibleForTesting;
 
 import static com.google.common.base.Preconditions.checkNotNull;
@@ -33,6 +34,9 @@ public class ParcelReception {
      * @return {@link ParcelReceipt package receipt}
      */
     public ParcelReceipt accept() {
+        RouteLookupRequest request = new RouteLookupRequest(parcel, sender);
+        Route route = LogisticsGateway.findRoute(request);
+
         Money price = estimatedPriceCalculator.calculate(parcel, sender);
         EstimatedDeliveryTime estimatedDeliveryTime = deliveryTime.estimate(parcel, sender);
         TrackNumber trackNumber = trackNumbers.issue(parcel);
