--- logistics/src/main/java/com/alevel/deliverit/postal/network/context/Context.java
@@ -1,4 +1,0 @@
-package com.alevel.deliverit.postal.network.context;
-
-public class Context implements SendingContext {
-}
