+++ model/src/main/java/com/alevel/deliverit/customers/ParcelId.java
@@ -6,7 +6,7 @@ import com.alevel.deliverit.entity.EntityId;
  * @author Vadym Mitin
  */
 public final class ParcelId extends EntityId<Long> {
-    protected ParcelId(Long value) {
+    public ParcelId(Long value) {
         super(value);
     }
 }
