+++ model/src/main/java/com/alevel/deliverit/customers/Sender.java
@@ -8,7 +8,7 @@ import com.alevel.deliverit.entity.Entity;
 public class Sender extends Entity<SenderId> {
     private final SenderProfile senderProfile;
 
-    protected Sender(SenderId id, SenderProfile senderProfile) {
+    public Sender(SenderId id, SenderProfile senderProfile) {
         super(id);
         this.senderProfile = senderProfile;
     }
