+++ model/src/main/java/com/alevel/deliverit/customers/SenderId.java
@@ -6,7 +6,7 @@ import com.alevel.deliverit.entity.EntityId;
  * @author Vadym Mitin
  */
 public final class SenderId extends EntityId<Long> {
-    protected SenderId(Long value) {
+    public SenderId(Long value) {
         super(value);
     }
 }
