+++ model/src/main/java/com/alevel/deliverit/logistics/postal/network/PostalUnit.java
@@ -1,4 +1,4 @@
-package com.alevel.deliverit.postal.network;
+package com.alevel.deliverit.logistics.postal.network;
 
 import com.google.common.collect.ImmutableSet;
 
@@ -7,7 +7,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-import static com.google.common.base.Preconditions.*;
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * A postal unit.
