+++ model/src/main/java/com/alevel/deliverit/logistics/postal/network/context/Context.java
@@ -1,0 +1,7 @@
+package com.alevel.deliverit.logistics.postal.network.context;
+
+/**
+ * @author Sergey Bogovesov
+ */
+public class Context implements SendingContext {
+}
